requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.20.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
Pillow>=9.5.0
facenet-pytorch>=2.5.3
mediapipe>=0.10.0
scikit-image>=0.21.0
ipykernel>=6.0.0
